% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{callduration}
\alias{callduration}
\title{callduration function
 Measures duration of individual vocalizations}
\usage{
callduration(xs)

}
\arguments{
\item{xs}{A single imported audio file or list of imported audio files}

}
\value{
A list of individual vocalization durations in seconds. Information for list is retrieved by duration function.

}
\description{
callduration function
 Measures duration of individual vocalizations

}
