% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{coolBlueHotRed}
\alias{coolBlueHotRed}
\title{coolBlueHotRed function
Create a vector of n contiguous colors}
\usage{
coolBlueHotRed(n, alpha = 1)

}
\arguments{
\item{n}{the number of colors (≥ 1) to be in the palette}

\item{alpha}{the alpha transparency, a number in [0,1]}

}
\value{
a default color palette for the kohonen heatmaps

}
\description{
coolBlueHotRed function
Create a vector of n contiguous colors

}
