% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{duration}
\alias{duration}
\title{duration function}
\usage{
duration(x)

}
\arguments{
\item{x}{An audio file}

}
\value{
Duration of individual vocalizations in seconds

}
\description{
Needed in Global Environment to activate call duration function

}
