% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{imp}
\alias{imp}
\title{imp function
Loads all wav files in specified working directory into named list}
\usage{
imp(mypath)

}
\arguments{
\item{mypath}{Directory in which files are stored}

}
\value{
Dataframe of audio files

}
\description{
imp function
Loads all wav files in specified working directory into named list

}
