% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{meancallduration}
\alias{meancallduration}
\title{meancallduration function}
\usage{
meancallduration(xs)

}
\arguments{
\item{xs}{A single imported audio file or list of imported audio files}

}
\value{
A list of mean vocalization durations in seconds. Information for list is retrieved by duration function.

}
\description{
Measures mean duration of vocalizations

}
