% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{meanduration}
\alias{meanduration}
\title{meanduration function}
\usage{
meanduration(x)

}
\arguments{
\item{x}{An imported audio file}

}
\value{
mean duration of calls in an audio file

}
\description{
Needed in Global Environment to activate meancallduration function

}
