% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{meansilence}
\alias{meansilence}
\title{meansilence function}
\usage{
meansilence(x)

}
\arguments{
\item{x}{An imported audio file}

}
\value{
Mean duration of silence per audio file

}
\description{
Needed in Global Environment to activate meansilenceduration function

}
