% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{meansilenceduration}
\alias{meansilenceduration}
\title{meansilenceduration function}
\usage{
meansilenceduration(xs)

}
\arguments{
\item{xs}{A single imported audio file or list of imported audio files}

}
\value{
A list of mean duration (in seconds) of silence between vocalizations. Information for list is retrieved by meansilence function.

}
\description{
Measures mean duration of silence between vocalizations

}
