% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{menzerath}
\alias{menzerath}
\title{menzerath function}
\usage{
menzerath(x)

}
\arguments{
\item{x}{A matrix created using the resulting lists from the seqsize function and either the meancallduration or meansilenceduration functions}

}
\value{
Two heatmaps, one for sequence size and one for constituent size, and the results of a Spearman's rank correlation test between the given variables

}
\description{
This function determines whether a dataset exhibits Menzerath's law by running the sequence and constituent variables through a Kohonen neural network. The output is a heatmap of each variables' distribution and a Spearman's rank correlation test on the relationship between the variables. A negative correlation suggests that the data exhibits Menzerath's law.

}
