% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{seqmeasure}
\alias{seqmeasure}
\title{seqmeasure function}
\usage{
seqmeasure(x)

}
\arguments{
\item{x}{An audio file}

}
\value{
Number of vocalizations per audio file

}
\description{
Needed in Global Environment to activate seqsize function

}
