% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{seqsize}
\alias{seqsize}
\title{seqsize function}
\usage{
seqsize(xs)

}
\arguments{
\item{xs}{A single imported audio file or a list of imported audio files}

}
\value{
A list of the number of vocalizations per audio file

}
\description{
Measures number of vocalizations per audio file

}
