% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{silence}
\alias{silence}
\title{silence function}
\usage{
silence(x)

}
\arguments{
\item{x}{An imported audio file}

}
\value{
Duration of individual periods of silence between vocalizations

}
\description{
Needed in Global Environment to activate silenceduration function

}
