% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{silenceduration}
\alias{silenceduration}
\title{silenceduration function}
\usage{
silenceduration(xs)

}
\arguments{
\item{xs}{A single imported audio file or list of imported audio files}

}
\value{
A list of individual periods of silence. Information for list is retrieved by silence function.

}
\description{
Measures duration of silence (in seconds) between vocalizations

}
