% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{silencemeasure}
\alias{silencemeasure}
\title{silencemeasure function}
\usage{
silencemeasure(x)

}
\arguments{
\item{x}{An audio file}

}
\value{
Number of periods of silence per audio file

}
\description{
Needed in Global Environment to activate silencesize function

}
