% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{silencesize}
\alias{silencesize}
\title{silencesize function}
\usage{
silencesize(xs)

}
\arguments{
\item{xs}{A single imported audio file or a list of imported audio files}

}
\value{
A list of the number of silence periods per audio file

}
\description{
Measures number of silences between vocalizations per audio file

}
