% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/menzerathfunctions.R
\name{visualize}
\alias{visualize}
\title{visualize function}
\usage{
visualize(x)

}
\arguments{
\item{x}{A single imported audio file or list of imported audio files}

}
\value{
plot of sound envelope with marked duration of calls and silences

}
\description{
Plots sound envelope. Marks calls and silences with duration.

}
